demo/requirements.txt
Flask==2.0.2
Turbo-Flask==0.8.0
gunicorn==20.1.0
invoke==1.7.1
requests==2.28.1
google-auth==2.3.2
google-crc32c==1.3.0
pycryptodome==3.15.0
google-cloud-storage==2.5.0
